refactor(myjobhunter): extract research sub-components, move commit to repo, add E2E
- Move db.commit() + db.refresh() out of company_research_service into
  company_research_repository.commit_with_sources_refresh() so the service
  layer has no direct transaction ownership
- Extract CompanyResearchPanelBody sub-components (NoResearchState, LoadingState,
  ReadyState, FailedState, FlagsSection, FlagList, SourcesList) into separate files
  under features/companies/research/ — one component per file
- Add e2e/company-research.spec.ts covering the no-research panel state and the
  error feedback path when Tavily is not configured in the test environment
